+++ env.d.ts
@@ -1,7 +1,5 @@
 /// <reference types="vite/client" />
 
-declare const __APP_VERSION__: string;
-
 declare module "@vanduo-oss/vd3/css";
 
 declare module "*.vue" {